examples/cloud-run/cloudrun-source-code/requirements.txt
functions-framework==3.5.0
google-cloud-logging==3.11.3
